refactor create: move relative code to a single impl
