+++ cli/src/print.rs
@@ -2,7 +2,7 @@ use crate::style;
 use crate::style::print_bold;
 use crate::style::print_green_bold;
 
-fn print_footer_message() {
+pub fn print_footer_message() {
     println!();
     println!(
         "📚 To learn more about mopro, visit: {}",
@@ -32,83 +32,5 @@ pub(crate) fn print_build_success_message() {
     println!();
     print_green_bold("Run the following command to create templates.".to_string());
     print_bold("   mopro create".to_string());
-    print_footer_message();
-}
-
-pub(crate) fn print_create_ios_success_message() {
-    print_green_bold("Template created successfully!".to_string());
-    println!();
-    print_green_bold("Next steps:".to_string());
-    println!();
-    print_green_bold("  You can now use the following command to open the app:".to_string());
-    println!();
-    print_bold("    open ios/MoproApp.xcodeproj".to_string());
-    println!();
-    print_green_bold("This will open the iOS project in Xcode.".to_string());
-    print_footer_message();
-}
-
-pub(crate) fn print_create_android_success_message() {
-    print_green_bold("Template created successfully!".to_string());
-    println!();
-    print_green_bold("Next steps:".to_string());
-    println!();
-    print_green_bold("  You can now use the following command to open the app:".to_string());
-    println!();
-    print_bold(r"    open android -a Android\ Studio ".to_string());
-    println!();
-    print_green_bold("This will open the Android project in Android Studio.".to_string());
-    print_footer_message();
-}
-
-pub(crate) fn print_create_web_success_message() {
-    print_green_bold("Template created successfully!".to_string());
-    println!();
-    print_green_bold("Next steps:".to_string());
-    println!();
-    print_green_bold("  You can now use the following command to start web server:".to_string());
-    println!();
-    print_bold(r"    cd web && yarn && yarn start".to_string());
-    println!();
-    print_green_bold("This will start a simple web server for your browser.".to_string());
-}
-
-pub(crate) fn print_create_flutter_success_message() {
-    print_green_bold("Flutter template created successfully!".to_string());
-    println!();
-    print_green_bold("Next steps:".to_string());
-    println!();
-    print_green_bold(
-        "  You can now use the following command to open the app in Android Studio:".to_string(),
-    );
-    println!();
-    print_bold(r"    open flutter-app -a Android\ Studio ".to_string());
-    println!();
-    print_green_bold("  or VS Code::".to_string());
-    println!();
-    print_bold(r"    code flutter-app".to_string());
-    println!();
-    print_green_bold(
-        "To learn more about setting up the Flutter app with mopro, visit https://zkmopro.org/docs/setup/flutter-setup/".to_string(),
-    );
-
-    print_footer_message();
-}
-
-pub(crate) fn print_create_react_native_success_message() {
-    print_green_bold("React Native template created successfully!".to_string());
-    println!();
-    print_green_bold("Next steps:".to_string());
-    println!();
-    print_green_bold(
-        "  You can now use the following command to open the app in VS Code:".to_string(),
-    );
-    println!();
-    print_bold(r"    code react-native-app".to_string());
-    println!();
-    print_green_bold(
-        "To learn more about setting up the React Native app with mopro, visit https://zkmopro.org/docs/setup/react-native-setup/".to_string(),
-    );
-
     print_footer_message();
 }
